Update README.md
FrontSteps is an information system that has been proposed to manage the registration of aid applicants and 
organizations that are helping them. Donors who want to contribute can then find information about the 
various organizations and their appeals for aid and make contributions which are either cash donations or 
goods. Donor information is not captured, however, the organizations would have to record the contributions 
received and the disbursements given.